refactor: separate delivery projection helpers
